requirements.txt
apache-airflow-providers-google
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-mongo==4.0.0
apache-airflow-providers-postgres==5.10.1
# jaydebeapi
# jpype1
numpy
pandas
psycopg2
pymongo
python-bsonjs
python-dateutil
